+++ src/mission/imap/+mag/+imap/+view/sftPlots.m
@@ -51,7 +51,7 @@ function figures = sftPlots(analysis, options)
                 m.crop([seconds(30), seconds(-30)]);
             end
 
-            views(end + 1) = mag.imap.view.PSD(m, Start = options.PSDStart, Duration = options.PSDDuration);
+            views(end + 1) = mag.graphics.view.PSD(m, Start = options.PSDStart, Duration = options.PSDDuration);
         end
     end